Fix parallel race in Sessy build metadata tests
Serialize REVISION/COMMITTED_AT file fixtures with flock so parallel
CI workers cannot delete each other's shared build files mid-setup.

+++ test/lib/sessy_build_metadata_test.rb
@@ -1,42 +1,14 @@
 require "test_helper"
 
 class SessyBuildMetadataTest < ActiveSupport::TestCase
-  setup do
-    @revision_path = Rails.root.join("REVISION")
-    @committed_at_path = Rails.root.join("COMMITTED_AT")
-    @had_revision = @revision_path.exist?
-    @had_committed_at = @committed_at_path.exist?
-    @previous_revision = @had_revision ? File.read(@revision_path) : nil
-    @previous_committed_at = @had_committed_at ? File.read(@committed_at_path) : nil
-  end
-
-  teardown do
-    restore_build_file(@revision_path, @had_revision, @previous_revision)
-    restore_build_file(@committed_at_path, @had_committed_at, @previous_committed_at)
-  end
-
   test "revision reads SESSY_GIT_SHA" do
     with_env("SESSY_GIT_SHA" => "deadbeef") do
       assert_equal "deadbeef", Sessy.revision
     end
   end
 
-  test "revision falls back to REVISION file" do
-    with_env("SESSY_GIT_SHA" => nil) do
-      File.write(@revision_path, "fromfile\n")
-      assert_equal "fromfile", Sessy.revision
-    end
-  end
-
   test "committed_at parses SESSY_GIT_COMMITTED_AT" do
     with_env("SESSY_GIT_COMMITTED_AT" => "2026-08-01T12:00:00Z") do
-      assert_equal Time.iso8601("2026-08-01T12:00:00Z"), Sessy.committed_at
-    end
-  end
-
-  test "committed_at falls back to COMMITTED_AT file" do
-    with_env("SESSY_GIT_COMMITTED_AT" => nil) do
-      File.write(@committed_at_path, "2026-08-01T12:00:00Z")
       assert_equal Time.iso8601("2026-08-01T12:00:00Z"), Sessy.committed_at
     end
   end
@@ -47,23 +19,79 @@ class SessyBuildMetadataTest < ActiveSupport::TestCase
     end
   end
 
+  test "revision falls back to REVISION file" do
+    with_build_files_locked do
+      with_env("SESSY_GIT_SHA" => nil) do
+        write_build_file("REVISION", "fromfile\n")
+        assert_equal "fromfile", Sessy.revision
+      end
+    end
+  end
+
+  test "committed_at falls back to COMMITTED_AT file" do
+    with_build_files_locked do
+      with_env("SESSY_GIT_COMMITTED_AT" => nil) do
+        write_build_file("COMMITTED_AT", "2026-08-01T12:00:00Z")
+        assert_equal Time.iso8601("2026-08-01T12:00:00Z"), Sessy.committed_at
+      end
+    end
+  end
+
   test "blank build files are treated as missing" do
-    with_env("SESSY_GIT_SHA" => nil, "SESSY_GIT_COMMITTED_AT" => nil) do
-      File.write(@revision_path, "   ")
-      File.write(@committed_at_path, "")
-      assert_nil Sessy.revision
-      assert_nil Sessy.committed_at
+    with_build_files_locked do
+      with_env("SESSY_GIT_SHA" => nil, "SESSY_GIT_COMMITTED_AT" => nil) do
+        write_build_file("REVISION", "   ")
+        write_build_file("COMMITTED_AT", "")
+        assert_nil Sessy.revision
+        assert_nil Sessy.committed_at
+      end
     end
   end
 
   private
 
-  def restore_build_file(path, had_file, contents)
-    if had_file
-      File.write(path, contents)
-    elsif path.exist?
-      File.delete(path)
+  def with_build_files_locked
+    FileUtils.mkdir_p(Rails.root.join("tmp"))
+    lock_path = Rails.root.join("tmp/sessy_build_metadata_test.lock")
+    File.open(lock_path, File::RDWR | File::CREAT, 0o644) do |lock|
+      lock.flock(File::LOCK_EX)
+      previous = snapshot_build_files
+      begin
+        yield
+      ensure
+        restore_build_files(previous)
+        lock.flock(File::LOCK_UN)
+      end
     end
+  end
+
+  def snapshot_build_files
+    {
+      "REVISION" => read_build_file_if_present("REVISION"),
+      "COMMITTED_AT" => read_build_file_if_present("COMMITTED_AT")
+    }
+  end
+
+  def restore_build_files(previous)
+    previous.each do |name, contents|
+      path = Rails.root.join(name)
+      if contents.nil?
+        File.delete(path) if path.exist?
+      else
+        File.write(path, contents)
+      end
+    end
+  end
+
+  def read_build_file_if_present(name)
+    path = Rails.root.join(name)
+    File.read(path) if path.exist?
+  rescue Errno::ENOENT
+    nil
+  end
+
+  def write_build_file(name, contents)
+    File.write(Rails.root.join(name), contents)
   end
 
   def with_env(vars)
